+++ src/main.js
@@ -1,12 +1,16 @@
 import Vue from "vue";
 import App from "./App.vue";
+import router from "@/router";
+import VueRouter from "vue-router";
 
 // Import Stylesheets
 import "@/assets/css/styles.css";
-import "@/assets/css/tailwind.css";
 
 Vue.config.productionTip = false;
 
+Vue.use(VueRouter);
+
 new Vue({
+  router,
   render: h => h(App),
 }).$mount("#app");