+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ### Added
 
 - Setup Vue2 project structure with tailwindcss and eslint
+- Added vue-router for page navigation
+- Added home page and 404 page with default layout including header and footer